Read from tileset and pick tileset according to Object
Now we have a working system to allow the system to read from what tileset we shall use tiles from. Its logic behind it is by using a second similiar although simplier grid called objectGrid. It is rn used to check if something is a wall or floor, but could be used for other stuff. Such as enemies, player, items etc.